Added a widget to list all libraries
This is my first attempt at really using the router.  There are a few issues still.  First,
you cannot jump directly to a route, you must follow hyperlinks.  This presumably has something
to do with the lack of fragments in the generated URLs.  I am guessing there is a setting
to use fragments and, if so, we should use it.

Another issue is that the header is still present.  Therefore, this listing of all libraries
is not really suitable for use with an iframe.