Got rid of matching component in state
I decided that this is probably something that is better handled in the Search component
(where it originally was).  I don't think many other components actually need to know
what the matching set is.

If multiple components were doing this calculation in parallel, I could see this as being
a useful optimization.  But since they aren't, there isn't any savings.  In fact, there are
times when we don't even care about matching components, so we even do some unnecessary
calculations.

Including it as a state also doesn't make sense since it is not really part of the state.
We don't need to know it to know the state of the system.  We can always derive the state
of the system from the list of available libraries in the index and the current search term.

So that is the rational behind this change.