Attempt to resolve security vulnerability in serialize-javascript
The problem here is that I don't use serialize-javascript directly and yet Yarn seems
to put it as a direct dependency when I do a 'yarn tree'.  I added an explicitly dependency
so that my immediate dependency in 'yarn tree' was a 'safe' version.  But this is actually
adding a dummy dependency just for the sake of silencing security errors (if it even works)
that involve indirect dependencies on serialize-javascript...which I can't control.